+++ supreme/src/androidMain/kotlin/at/asitplus/signum/supreme/sign/VerifierImpl.kt
@@ -9,9 +9,6 @@ import at.asitplus.signum.indispensable.jcaAlgorithmComponent
 import at.asitplus.signum.indispensable.jcaSignatureBytes
 import at.asitplus.signum.supreme.dsl.DSL
 import at.asitplus.signum.supreme.UnsupportedCryptoException
-import at.asitplus.signum.supreme.sign.InvalidSignature
-import at.asitplus.signum.supreme.sign.PlatformVerifierConfiguration
-import at.asitplus.signum.supreme.sign.SignatureInput
 import at.asitplus.wrapping
 import java.security.Signature
 
@@ -70,7 +67,7 @@ private fun getRSAInstance(alg: SignatureAlgorithm.RSA, config: PlatformVerifier
 
 @Throws(UnsupportedCryptoException::class)
 internal actual fun checkAlgorithmKeyCombinationSupportedByRSAPlatformVerifier
-            (signatureAlgorithm: SignatureAlgorithm.RSA, publicKey: CryptoPublicKey.Rsa,
+            (signatureAlgorithm: SignatureAlgorithm.RSA, publicKey: CryptoPublicKey.RSA,
              config: PlatformVerifierConfiguration)
 {
     wrapping(asA=::UnsupportedCryptoException) {
@@ -81,7 +78,7 @@ internal actual fun checkAlgorithmKeyCombinationSupportedByRSAPlatformVerifier
 
 @JvmSynthetic
 internal actual fun verifyRSAImpl
-    (signatureAlgorithm: SignatureAlgorithm.RSA, publicKey: CryptoPublicKey.Rsa,
+    (signatureAlgorithm: SignatureAlgorithm.RSA, publicKey: CryptoPublicKey.RSA,
      data: SignatureInput, signature: CryptoSignature.RSAorHMAC,
      config: PlatformVerifierConfiguration)
 {