+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.0
 
+### 3.10.0 NEXT
+
 ### 3.9.0 (Supreme 0.4.0)
 
 * Move `Attestation` from Supreme to Indispensable